revert settings changes
