+++ CMakeLists.txt
@@ -166,7 +166,6 @@ set(IMGUI_SOURCES
 # Utility library
 add_library(util STATIC
   util/settings.cpp
-  util/settings_file_manager.cpp
   util/terminal.cpp
   util/close_popper.cpp
   util/welcome.cpp
